Fix autowiring error for GrouperService
Add new unit test for grouperApiService
